Tweaks to match Adrian's specs, also switched to an index.jsx file for each component
