+++ src/components/Byline/Byline.css
@@ -1,4 +1,5 @@
-.Byline{
-	font-size: 1.15em;
-	margin: 0;
+.Byline {
+    /*	font-size: 1.15em;*/
+    color: #f8f800;
+    margin: 0;
 }
