+++ src/components/BylineContainer/BylineContainer.css
@@ -1,3 +1,7 @@
+.BylineContainer{
+	font-size: 16px;
+}
+
 @media (min-width: 1024px) {
     .BylineContainer {
         margin-top: 2em;
