+++ src/components/Headline/Headline.css
@@ -1,7 +1,8 @@
 .Headline {
-    font-size: 1.5em;
-    font-weight: bold;
-    margin: 1em 0 .25em;
+    font-size: 30px;
+    font-weight: normal;
+/*    margin: 1em 0 .25em;*/
+    font-family: 'Suez One', serif;
 }
 
 @media (max-width: 1023px) {
