+++ src/components/LeadPhotoContainer/index.jsx
