+++ src/components/LeftYellowBorder/LeftYellowBorder.css
@@ -1,4 +1,4 @@
 .LeftYellowBorder{
-	border-left: 8px solid #ff9900;
+	border-left: 8px solid #FCB001;
 	padding-left: .85em;
 }
