+++ src/components/PhotoEssay/PhotoEssay.jsx
@@ -1,5 +1,3 @@
-// @flow
-
 import React from 'react';
 import _ from 'lodash';
 import ReactVisiblitySensor from 'react-visibility-sensor';
