+++ src/components/PhotoEssayContainer/PhotoEssayContainer.jsx
@@ -1,5 +1,3 @@
-// @flow
-
 import React, { Component } from 'react';
 
 // grommet components
