+++ src/components/PhotoInfo/PhotoInfo.css
@@ -3,6 +3,12 @@
 	width: 90%;
 }
 
+@media (max-width: 767px){
+	.StoryTop .PhotoInfo{
+		display: none;
+	}
+}
+
 
 
 @media only screen and (min-width: 1024px){
