+++ src/components/StoryMeta/StoryMeta.css
@@ -1,6 +1,9 @@
 .StoryMeta{
-	margin: 0 auto;
-	width: 90%;
+/*	margin: 0 auto;
+	width: 90%;*/
+	background: black;
+	color: #FDFEFE;
+	padding-bottom: 1em;
 }
 
 @media only screen and (min-width: 768px){
