+++ src/components/StoryTop/index.jsx
@@ -2,8 +2,9 @@
 
 import React, {Component} from 'react';
 
-import LeadPhotoContainer from '../LeadPhotoContainer/LeadPhotoContainer';
-import StoryMeta from '../StoryMeta/StoryMeta';
+import LeadPhotoContainer from '../LeadPhotoContainer';
+import StoryMeta from '../StoryMeta';
+import StorySeriesBlock from '../StorySeriesBlock';
 
 import './StoryTop.css';
 
@@ -16,7 +17,7 @@ class StoryTop extends Component {
 		return(
 			<div className='StoryTop'>
 				<LeadPhotoContainer data={this.props.data.image} />
-				<StoryMeta meta={this.props.data.meta}/>
+				<StoryMeta meta={this.props.data.meta}/> 
 			</div>
 		)
 	}
